Move raft-related code to replicated state machine (rsm) namespace
