+++ src/io/rsm/rsm.hpp
@@ -1,0 +1,26 @@
+// Copyright 2022 Memgraph Ltd.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
+// License, and you may not use this file except in compliance with the Business Source License.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+#pragma once
+
+/// Replicated State Machine-related code.
+namespace memgraph::io::rsm {
+
+/*
+template <typename T>
+concept ReplicatedStateMachine = true;
+requires(T a, uint8_t *ptr, size_t len) {
+  { a.Serialize() } -> std::same_as<std::vector<uint8_t>>;
+  { T::Deserialize(ptr, len) } -> std::same_as<T>;
+};
+*/
+
+};
